% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-version.R
\name{add_feature_version}
\alias{add_feature_version}
\title{Add a feature version number to an R package}
\usage{
add_feature_version(pkg_path = ".")
}
\arguments{
\item{pkg_path}{The path to the package. Defaults to the current working}
}
\value{
An updated package DESCRIPTION, with the feature version number added.
}
\description{
Adds a feature version number to an R package, namely by adding a
dash and the Unix epoch at the time the command was ran to the build version.
}
\author{
Waldir Leoncio
}
