% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert-breakpoint.R
\name{insert_breakpoint}
\alias{insert_breakpoint}
\title{Insert breakpoint in a function}
\usage{
insert_breakpoint(fun_name)
}
\arguments{
\item{fun_name}{name of the function}
}
\value{
A traced function at the selected point
}
\description{
Interactively inserts a breakpoint on a function under development
}
\author{
Waldir Leoncio
}
