% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-version.R
\name{remove_feature_version}
\alias{remove_feature_version}
\title{Removes the feature version number from an R package}
\usage{
remove_feature_version(pkg_path = ".")
}
\arguments{
\item{pkg_path}{The path to the package. Defaults to the current working}
}
\value{
An updated package DESCRIPTION, with the feature version number removed.
}
\description{
Reverts the work done by \code{\link[=add_feature_version]{add_feature_version()}} by removing the
feature version number from the package's DESCRIPTION file.
}
\author{
Waldir Leoncio
}
